services/indexer/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
pydantic-settings==2.4.0
beautifulsoup4==4.12.3
pypdf==4.3.1
python-docx==1.1.2
markdown-it-py==3.0.0
boto3==1.34.162
kafka-python==2.0.2
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
qdrant-client==1.10.1
opensearch-py==2.6.0
pytest==8.3.2
httpx==0.27.0
